+++ src/Common/index.js
@@ -3,3 +3,4 @@ export { default as DeleteAlertModal } from './DeleteAlertModal/DeleteAlertModal
 export { default as ErrorPage } from './ErrorPage/ErrorPage.component';
 export { default as SnackBar } from './SnackBar/SnackBar.component';
 export { default as Spinner } from './Spinner/Spinner.component';
+export { default as NavBar } from './NavBar/NavBar.component';
