+++ src/Components/Games/GameForm.component.jsx
@@ -2,8 +2,8 @@ import { useState, useEffect, useContext } from 'react';
 import { Button, Modal, Form, Row, Col, InputGroup, Badge } from 'react-bootstrap';
 import proptypes from 'prop-types';
 import { XCircle } from "react-bootstrap-icons";
-import { AppState } from "../Config/store/state";
-import { gameObjectSanitizer } from '../utils/requestSanitizer';
+import { AppState } from "../../Config/store/state";
+import { gameObjectSanitizer } from '../../utils/requestSanitizer';
 import classes from './Games.module.css';
 
 const GAME_DEFAULT = {
