+++ src/Config/Routes.jsx
@@ -1,13 +1,13 @@
 import { createBrowserRouter } from "react-router-dom";
-import Dashboard from "../Home/Dashboard.component";
+import Dashboard from "../Components/Home/Dashboard.component";
 import BrandsContainer from "../Admin/Brand/Brands.container";
 import GenresContainer from "../Admin/Genre/Genres.container";
 import Layout from '../Layout/Layout.component';
 import ErrorBoundaryLayout from "../Layout/ErrorBoundaryLayout";
 import { PageNotFound } from "../Common/ErrorPage";
-import Brands from "../Brands/Brands.container";
-import Consoles from "../Consoles/Consoles.container";
-import Games from "../Games/Games.container";
+import Brands from "../Components/Brands/Brands.container";
+import Consoles from "../Components/Consoles/Consoles.container";
+import Games from "../Components/Games/Games.container";
 
 
 const router = createBrowserRouter([
