+++ src/Layout/Layout.component.jsx
@@ -1,8 +1,7 @@
 import proptypes from 'prop-types';
 import { Outlet } from 'react-router-dom'
 import Content from './Content/Content.component';
-import NavBar from '../Navigation/NavBar/NavBar.component';
-import SnackBar from '../Common/SnackBar/SnackBar.component';
+import { NavBar, SnackBar } from '../Common';
 
 const Layout = ({children}) => (
   <>
